find_value: Make truncation to int explicit
warning: signed and unsigned type in conditional expression [-Wsign-compare]

+++ utils/dump.c
@@ -547,7 +547,8 @@ static int
 find_value (ElfW(Addr) base, void *start, size_t size, int what)
 {
     const ElfW(Dyn) *entry = find_dyn( base, start, size, what );
-    return entry ? entry->d_un.d_val : -1;
+    // TODO: what if it doesn't fit in an int?
+    return entry ? (int) entry->d_un.d_val : -1;
 }
 
 static ElfW(Addr)
