+++ utils/print-libstubs.c
@@ -209,7 +209,8 @@ static int
 find_value (ElfW(Addr) base, void *start, int what)
 {
     const ElfW(Dyn) *entry = find_dyn( base, start, what );
-    return entry ? entry->d_un.d_val : -1;
+    // TODO: what if it doesn't fit in an int?
+    return entry ? (int) entry->d_un.d_val : -1;
 }
 
 static ElfW(Addr)
